+++ core/src/main/java/org/apache/gravitino/hook/DispatcherHookHelper.java
@@ -1,0 +1,35 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.gravitino.hook;
+
+import java.lang.reflect.Proxy;
+
+/** The class is a helper class of dispatcher hooks */
+public class DispatcherHookHelper {
+
+  private DispatcherHookHelper() {}
+
+  public static <T> T installHooks(T dispatcher, DispatcherHooks hooks) {
+    return (T)
+        Proxy.newProxyInstance(
+            dispatcher.getClass().getClassLoader(),
+            dispatcher.getClass().getInterfaces(),
+            new DispatcherHookProxy<T>(dispatcher, hooks));
+  }
+}
